JBTM-3838 Ensure that the checking for destroy state does not prevent the removeConnection from being called

+++ ArjunaJTS/jtax/classes/com/arjuna/ats/internal/jta/resources/jts/orbspecific/XAResourceRecord.java
@@ -431,9 +431,8 @@ public class XAResourceRecord extends com.arjuna.ArjunaOTS.OTSAbstractRecordPOA
 				finally
 				{
 				    _phaseTwoStarted = true;
-					if (_prepared)
-						if (destroyState)
-							destroyState();
+					if (_prepared && destroyState)
+						destroyState();
 					else
 						removeConnection();
 				}
